feat:【PAC模板】流水线模板支持PAC特性 #11414

+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/engine/dao/template/TemplateInstanceBaseDao.kt
@@ -218,7 +218,7 @@ class TemplateInstanceBaseDao {
             baseId = id,
             projectId = projectId,
             templateId = templateId,
-            templateVersion = templateVersion.toInt(),
+            templateVersion = templateVersion.toLong(),
             useTemplateSetting = useTemplateSettingsFlag,
             totalItemNum = totalItemNum,
             successItemNum = successItemNum,
